+++ packages/react/src/components/MenuButton/src/components/MenuButtonAccordion/index.ts
@@ -1,0 +1,4 @@
+import { MenuButtonAccordion } from "./MenuButtonAccordion";
+
+export { MenuButtonAccordion } from "./MenuButtonAccordion";
+export default MenuButtonAccordion;